pyproject.toml
[tool.poetry]
name = "drip-search"
version = "0.1.0"
description = ""
authors = ["senya <arseniyohar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
numpy = "^2.1.3"
torch = "^2.5.1"
datasets = {extras = ["text"], version = "^3.1.0"}
huggingface-hub = "^0.26.3"
sentence-transformers = "^3.3.1"
pytest = "^8.3.4"
python-dotenv = "^1.0.1"
groq = "^0.13.0"
streamlit = "^1.40.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
